requirements.txt
flask==3.0.3
flask-pymongo==2.3.0
pymongo==4.7.1
scikit-learn==1.6.1
nltk==3.8.1
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
Werkzeug==3.0.3